fix: `kubeconfig` argument is now honored
Before, the `kubeconfig` argument was specified, but not really used.
This has now been fixed. There is now a `PersistentPreRun` hook in place
in the cobara root command, which will handle the initialization of the
kubernetes client. When this hook is executed, the `kubeconfig` and
`masterURL` arguments are filled and are now handled when creating a new
kubernetes client.

Signed-off-by: Patrick Pichler <git@patrickpichler.dev>